Module type Mimic_happy_eyeballs.S

type t

type flow

val happy_eyeballs : t Mimic.value

val resolve : t -> ?aaaa_timeout:int64 -> ?connect_delay:int64 -> ?connect_timeout:int64 -> ?resolve_timeout:int64 -> ?resolve_retries:int -> string -> int list -> ((Ipaddr.t * int) * flow, [> `Msg of string ]) Stdlib.result Lwt.t

resolve happy_eyeballs domain_name ports tries to connect the user with the given endpoint. The domain-name can be an IP address or a real domain-name. resolve tries to resolve via a DNS request the given domain_name if it's a real domain-name and it tries to initiate a TCP/IP connection with the destination.

If it's a success, it returns the resource and the user is able to read or write via this resource.